🐛 Return 410 from remaining cart-dependent composers when slice is reaped
The summary GET already 410s on a missing cart slice; this extends the
same treatment to every other cart-dependent composer:

- SummarySubmitComposer (POST /buy/summary/{orderId}) - defensive 410
  in case the slice is reaped between summary load and submit click
  so we don't silently submit an empty order
- ShippingScreenCartComposer (GET /buy/shipping/{orderId})
- CartSummaryComposer (GET /buy/address/{orderId} and /buy/payment/{orderId})
- ShoppingCartComposer (GET /cart/{orderId})

ShoppingCartAddItemComposer keeps its empty-slice fallback - that one
legitimately bootstraps a fresh cart on the customer's first add.

A shared ExpiredCartNotice.svelte renders the "session expired" view,
wired into the cart, address, shipping, payment, and summary pages.
The summary page also handles the placeOrder catch path. Cart-count
refresh in the layout clears the orderId on 410 so a dead pointer
doesn't keep haunting subsequent page loads.
