+++ src/Catalog.ServiceComposition/Cart/ShoppingCartComposer.cs
@@ -19,8 +19,17 @@ public class ShoppingCartComposer(IWorkflowStore workflow, CatalogDbContext dbCo
         var vm = request.GetComposedResponseModel();
         var ct = request.HttpContext.RequestAborted;
 
-        var cart = await workflow.Read<CartSlice>(orderId, CartWorkflowSlice.Key, ct)
-                   ?? CartSlice.Empty;
+        // Returning an empty cart on a missing slice silently hides the
+        // fact that WorkflowCleanupHostedService reaped what the
+        // customer had. 410 lets the SPA show an explicit
+        // "your cart was cleared" view and reset the orderId so the
+        // next interaction starts fresh.
+        var cart = await workflow.Read<CartSlice>(orderId, CartWorkflowSlice.Key, ct);
+        if (cart is null)
+        {
+            request.SetActionResult(new StatusCodeResult(StatusCodes.Status410Gone));
+            return;
+        }
 
         var productIds = cart.Items.Select(s => s.ProductId).ToList();
         var products = await dbContext.Products
