+++ src/website/src/lib/stores/cart.js
@@ -1,5 +1,6 @@
 import { writable } from 'svelte/store';
 import { gateway } from '$lib/api/gateway.js';
+import { orderId as orderIdStore } from '$lib/stores/orderId.js';
 
 export const cartItemCount = writable(0);
 
@@ -15,7 +16,11 @@ export async function refreshCartCount(orderId) {
       0
     );
     cartItemCount.set(count);
-  } catch {
+  } catch (e) {
     cartItemCount.set(0);
+    // 410 from the cart composer means the workflow store reaped the
+    // slice this orderId points at - it's a dead pointer, so clear it
+    // and let the next add-to-cart create a fresh one.
+    if (e?.status === 410) orderIdStore.clear();
   }
 }
